rename LedgerView to StagedLedger, define new ledger view to resolve commits
