+++ fluree-db-transact/src/commit.rs
@@ -9,7 +9,7 @@ use crate::raw_txn_upload::PendingRawTxnUpload;
 use chrono::Utc;
 use fluree_db_binary_index::BinaryIndexStore;
 use fluree_db_core::{ContentId, ContentKind, ContentStore, DictNovelty, Flake, TXN_META_GRAPH_ID};
-use fluree_db_ledger::{IndexConfig, LedgerState, LedgerView};
+use fluree_db_ledger::{IndexConfig, LedgerState, StagedLedger};
 use fluree_db_nameservice::{CasResult, NameService, RefKind, RefPublisher, RefValue};
 use fluree_db_novelty::{generate_commit_flakes, stamp_graph_on_commit_flakes};
 use fluree_db_novelty::{Commit, SigningKey, TxnMetaEntry, TxnMetaValue, TxnSignature};
@@ -287,7 +287,7 @@ impl CommitOpts {
 ///
 /// A tuple of (CommitReceipt, new LedgerState)
 pub async fn commit<C, N>(
-    view: LedgerView,
+    view: StagedLedger,
     mut ns_registry: NamespaceRegistry,
     content_store: &C,
     nameservice: &N,
